Switching between frames
Defined the frames in the controller in a dictionary. 
Set the variables to call frame functions. 
Defined a new function to use tkraise() to raise frames on a button click.
Use a lambda function to call the function to show frames within a frame by calling the frame.
Changed parameters to the defined Timer and Settings classes to accept frames in the init method.

Added a settings / timer button within the frames to switch between the two frames.

Added to buttons to grid for settings and timer frame, to call functions to show frames.

Made additional configurations to tighten the look of the frames